requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-dotenv
google-generativeai
mysql-connector-python
